arm: juno: Enable OF_CONTROL
The Arm Juno board was still somewhat stuck in "hardcoded land", even
though there are stable DTs around, and one happens to actually be on
the memory mapped NOR flash.

Enable the configuration options to let the board use OF_CONTROL, and
add a routine to find the address of the "board.dtb" partition in NOR
flash, to use that for U-Boot's own purposes.
This can also passed on via $fdtcontroladdr to any kernel, removing the
need to acutally load a device tree.

Since the "afs" flash routines require flash_init() to be called before
being usable, and this is done much later in the boot process, we
introduce a stripped-down partition finder routine in vexpress64.c, to
scan the NOR flash partitions for the DT partition. This location is
then used for U-Boot to find and probe devices.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>