qemu: Drop ARCH_SUPPORT_TFABOOT
CONFIG_ARCH_SUPPORT_TFABOOT seems to be a guard option to enable various
platfrom specific hacks, when U-Boot is run under TF-A.
Now that the QEMU port does not need to differentiate between secure
vs. non-secure anymore (this is taken care of by the DTB), there is
no need for a build-time option anymore, so remove it.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>