vexpress: Disable MMC
The PL180 MMC controller driver is DM capable, but also requires a DM
GPIO driver for the card select GPIO.

For the VExpress boards this GPIO is handled via the sysregs registers,
for which we don't have any driver yet. Plus the devicetree description
of those registers is somewhat special, which complicates the matter.

Most people prefer Ethernet over the MMC controller anyway, so let's
just disable the MMC feature for now, to make the platform compile
without deprecation warnings.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>