vexpress: add ARCH_VEXPRESS, use board choice in Kconfig
Factor out the common parts of the Kconfig into an ARCH_VEXPRESS symbol,
then let the three boards build on that.
Since we can only build for one of these, let's make this a Kconfig
"choice".
This also uses a common config name, so we drop the three non-redundant
per-board header files.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>