vexpress: Move SMP spin address into shared config
We support non-secure boot (and thus SMP support) only for the VExpress
TC2, but this does not justify having separate config files for each
board.

Move this definiton, which is the only relevant difference left between
the three VExpress boards, to the shared header file.

This allows us to drop those separate files later.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>